+++ pkg/kclient/kclient_test.go
@@ -2,6 +2,7 @@ package kclient
 
 import (
 	"fmt"
+	"github.com/golang/mock/gomock"
 	"runtime"
 	"testing"
 
@@ -103,5 +104,49 @@ func TestDelete(t *testing.T) {
 	err := fkclient.Delete(selectors, false)
 	if err != nil {
 		t.Errorf("Expected no error, got %s", err)
+	}
+}
+
+func TestKClient_CheckDefaultProject(t *testing.T) {
+	tests := []struct {
+		testName    string
+		supported   bool
+		wantErr     bool
+		projectName string
+		supportErr  error
+	}{
+		{
+			testName:    "Case 0: CheckDefaultProject returns no error",
+			projectName: "myproject",
+			supported:   true,
+			wantErr:     false,
+			supportErr:  nil,
+		},
+		{
+			testName:    "Case 1: CheckDefaultProject returns error on using 'default' project name in OC",
+			projectName: "default",
+			supported:   true,
+			wantErr:     true,
+			supportErr:  nil,
+		},
+		{testName: "Case 2: CheckDefaultProject returns error on checking if the project resource is supported",
+			projectName: "myproject",
+			supported:   false,
+			wantErr:     true,
+			supportErr:  fmt.Errorf("some error while checking project support"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.testName, func(t *testing.T) {
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+			mockKClient := NewMockClientInterface(ctrl)
+			mockKClient.EXPECT().IsProjectSupported().Return(tt.supported, tt.supportErr).AnyTimes()
+			//  TODO: This wil not work (pvala)
+			got := mockKClient.CheckDefaultProject(tt.projectName)
+			if (!tt.wantErr && got != nil) || (tt.wantErr && got == nil) {
+				t.Errorf("got==nil: %v; wantErr: %v", got == nil, tt.wantErr)
+			}
+		})
 	}
 }
